I somehow failed to run all the tests and therefore didn’t see that this one started failing.
I don’t think there was ever a use case for FromDoc exceptions being passed through unmodified (and they kinda weren’t before, if your FromDoc threw a FormatException).  I think it was just a case of a unit test baking in an assumption.  And now we’re baking in a different assumption!